Extract IAM content detail resolution
